Basic AR.js project with Webpack
Set up a basic AR.js project and used webpack to bundle the javascript. Chose to build the AR.js project on top Three.js instead of A-Frame because there seems to be more support and a larger community which may help us down the line.

One downside of AR.js is that a library they use in order to render the objects onto the camera view can only be accessed from a secure origin i.e HTTPS meaning we will need a SSL certificate in the future. This means you can't run the application on your computer and access it via your phone without the cert. But, it can be tested locally on the same machine hosting it.