Added Markers to Map
Added the markers to the map view. For now I thought he McGill crest would look best but it can be changed in the future. Still a lot to do from handling the click events to the markers to filtering the results in the search.

Also, commented out the bottom sheet for now because the map nor search bear received click events when it is present although it is closed. Weird bug