More Map Features
Changed the look of the search bar so its more seameless. Now, when you click on a search result it selects the marker on the map and updates the location. Additionally, when you select either the marker or the search result it populates the bottom sheet and appears. However, theres still a bug with click events outside of the bottomsheet not registering.