Additional Changes Related to Layout on Android
Changed the height of the bottom sheet so it is more visible on Android, particularly the Google Pixel 2. The height will not be perfect on every device, but a new version of the bottom sheet which is in alpha dynamically sets the bottom sheet to the height of the content which we can use when it's released. Also, hid the bottomsheet when searching because the behaviour on the Pixel is not the same on Apple and it clutters the screen.