Image Upload and Fetch Logic and formatting
Most of the lines in this commit are due to formatting. The rest are related to the logic needed to fetch images from the db and also upload images to the db along with a schedule modal which can be used to view the current schedule if one is uploaded. The logic for deleting a user is still not in place and still need some logic for updating an image in the db.