Modified image request type and some logic for image uploads
Changed the POST requests to PUT requests to reflect the changes made on the backend. Also checked to see if the PUT request fails we can assume that it was due to an image upload size error. Although, this should be handled on the backend and a different status code should be returned it is good for now.